# backend/docs/swagger.yaml
definitions:
...
  entity.UserRole:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Admin
    - Organizer
    - Participant
...
  user_dto.Auth:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  user_dto.SignUp:
    properties:
      login:
        type: string
      name:
        type: string
      password:
        type: string
      user_role:
        $ref: '#/definitions/entity.UserRole'
    type: object
  virtual_room_dto.CreateVirtualRoom:
    properties:
      capacity:
        type: integer
      room_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /users/{login}:
    delete:
      description: Удаляет пользователя по его логину
      parameters:
      - description: Логин пользователя
        in: path
        name: login
        required: true
        type: string
      responses:
        "200":
          description: Успешное удаление
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Удалить пользователя
      tags:
      - Users
    get:
      description: Возвращает данные пользователя по его логину
      parameters:
      - description: Логин пользователя
        in: path
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о пользователе
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить пользователя
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Обновляет данные пользователя по его логину
      parameters:
      - description: Логин пользователя
        in: path
        name: login
        required: true
        type: string
      - description: Обновляемые данные
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Обновить пользователя
      tags:
      - Users
  /users/all:
    get:
      description: Возвращает список всех зарегистрированных пользователей
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Список пользователей
      tags:
      - Users
  /users/auth:
    post:
      consumes:
      - application/json
      description: Проверяет учетные данные пользователя и возвращает токен
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user_dto.Auth'
      produces:
      - application/json
      responses:
        "200":
          description: Токен аутентификации
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Аутентификация пользователя
      tags:
      - Users
  /users/signUp:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user_dto.SignUp'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная регистрация
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Регистрация пользователя
      tags:
      - Users
  /virtualRooms/{name}:
    delete:
      description: Удаляет виртуальную комнату по названию
      parameters:
      - description: Название комнаты
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: Название удалённой комнаты
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Удалить виртуальную комнату
      tags:
      - VirtualRooms
    get:
      description: Возвращает данные виртуальной комнаты по её названию
      parameters:
      - description: Название комнаты
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о комнате
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить виртуальную комнату
      tags:
      - VirtualRooms
  /virtualRooms/all:
    get:
      description: Возвращает список всех виртуальных комнат
      produces:
      - application/json
      responses:
        "200":
          description: Список виртуальных комнат
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Список виртуальных комнат
      tags:
      - VirtualRooms
  /virtualRooms/create:
    post:
      consumes:
      - application/json
      description: Добавляет новую виртуальную комнату в систему
      parameters:
      - description: Данные комнаты
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/virtual_room_dto.CreateVirtualRoom'
      produces:
      - application/json
      responses:
        "200":
          description: Название комнаты
          schema:
            type: string
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Создать виртуальную комнату
      tags:
      - VirtualRooms
  /virtualRooms/update/{name}:
    put:
      consumes:
      - application/json
      description: Обновляет данные виртуальной комнаты по названию
      parameters:
      - description: Название комнаты
        in: path
        name: name
        required: true
        type: string
      - description: Обновляемые данные
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённые данные комнаты
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Обновить виртуальную комнату
      tags:
      - VirtualRooms
swagger: "2.0"